add unittests for encoder, decoder and transformer
